requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
openai==1.3.0
Pillow==10.0.1
qrcode==7.4.2
schedule==1.2.0
flask==2.3.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
feedparser==6.0.10
newspaper3k==0.2.8
transformers==4.35.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
